Cleanup & removal of old menu code
