+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+ToolbarDelegate.swift
@@ -448,9 +448,9 @@ extension BrowserViewController: ToolbarDelegate {
         }
     }
     
-    func featuresMenuSection(_ menuController: NewMenuController) -> some View {
+    func featuresMenuSection(_ menuController: MenuViewController) -> some View {
         VStack(spacing: 0) {
-            VPNMenuButton(icon: #imageLiteral(resourceName: "vpn_menu_icon").template, title: "Brave VPN", vpnProductInfo: self.vpnProductInfo) { vc in
+            VPNMenuButton(vpnProductInfo: self.vpnProductInfo) { vc in
                 self.dismiss(animated: true) {
                     self.present(vc, animated: true)
                 }
@@ -458,7 +458,7 @@ extension BrowserViewController: ToolbarDelegate {
         }
     }
     
-    func destinationMenuSection(_ menuController: NewMenuController) -> some View {
+    func destinationMenuSection(_ menuController: MenuViewController) -> some View {
         VStack(spacing: 0) {
             MenuItemButton(icon: #imageLiteral(resourceName: "menu_bookmarks").template, title: Strings.bookmarksMenuItem) {
                 let vc = BookmarksViewController(folder: Bookmarkv2.lastVisitedFolder(), isPrivateBrowsing: PrivateBrowsingManager.shared.isPrivateBrowsing)
@@ -472,6 +472,7 @@ extension BrowserViewController: ToolbarDelegate {
             }
             MenuItemButton(icon: #imageLiteral(resourceName: "menu-downloads").template, title: Strings.downloadsMenuItem) {
                 let vc = DownloadsPanel(profile: self.profile)
+                vc.applyTheme(Theme.of(nil))
                 menuController.pushViewController(vc, animated: true)
             }
             MenuItemButton(icon: #imageLiteral(resourceName: "menu-settings").template, title: Strings.settingsMenuItem) {
@@ -506,7 +507,7 @@ extension BrowserViewController: ToolbarDelegate {
         }
     }
     
-    func activitiesMenuSection(_ menuController: NewMenuController, tabURL: URL, activities: [UIActivity]) -> some View {
+    func activitiesMenuSection(_ menuController: MenuViewController, tabURL: URL, activities: [UIActivity]) -> some View {
         VStack(alignment: .leading, spacing: 0) {
             MenuTabDetailsView(tab: tabManager.selectedTab, url: tabURL)
             VStack(spacing: 0) {
@@ -539,7 +540,7 @@ extension BrowserViewController: ToolbarDelegate {
         if let url = selectedTabURL, let tab = tabManager.selectedTab {
             activities = shareActivities(for: url, tab: tab, sourceView: view, sourceRect: self.view.convert(self.topToolbar.menuButton.frame, from: self.topToolbar.menuButton.superview), arrowDirection: .up)
         }
-        let menuController = NewMenuController(content: { menuController in
+        let menuController = MenuViewController(content: { menuController in
             VStack(spacing: 6) {
                 featuresMenuSection(menuController)
                 Divider()
